Clean up func_fw_decrypt_run_c
This simply calls fw_decrypt_atj2127 in a loop over all sectors.
Intriguingly, though, it reinitialises its state every time it is
called, and it's only ever asked to decrypt a maximum of 2k of data.

I realise that the whole scheme is very weak, cryptographically, but
this part is particularly odd, since doing the right thing wouldn't have
been any extra work.